MacGUI: Forgot a file in my previous commit.
git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@1394 b64f7644-9d1e-0410-96f1-a4d463321fa5

+++ macosx/HBQueueController.h
@@ -164,8 +164,8 @@ BOOL                        fIsDragging;
     NSMutableArray               *fJobs;   // array of HBJob
     NSMutableAttributedString    *fDescription;
     BOOL                         fNeedsDescription;
-    float                        fLastDescriptionHeight;
-    float                        fLastDescriptionWidth;
+    CGFloat                        fLastDescriptionHeight;
+    CGFloat                        fLastDescriptionWidth;
     HBQueueJobGroupStatus        fStatus;
     NSString                     *fPresetName;
 }
@@ -191,8 +191,8 @@ BOOL                        fIsDragging;
 // Creating a description
 - (void)               setNeedsDescription: (BOOL)flag;
 - (NSMutableAttributedString *) attributedDescription;
-- (float)              heightOfDescriptionForWidth:(float)width;
-- (float)              lastDescriptionHeight;
+- (CGFloat)              heightOfDescriptionForWidth:(CGFloat)width;
+- (CGFloat)              lastDescriptionHeight;
 
 @end
